Initial commit
This is the base for the Qt / QGraphicsScene rewrite of picmi. The
chaotic first phase of development should be over and therefore we can
finally get started on a decent version history.

Most old functionality has been reimplemented. Still missing are
highscore handling, layout and graphics improvements and generally lots
of polish and refactoring.