Work around an apparent bug with old gcc versions.
Building with gcc 4.2.1 failed because QList::append() was called
without a template parameter being passed to it.

Since it works fine with gcc 4.6.3 and clang 3.1, it seems to be a bug
with old gcc releases. Work around it by explicitly passing the
template parameter to QList.

CCMAIL: jakob.gruber@gmail.com