Search the correct locations in QStandardPath::locate
Passing both an organization name and application name to
QCoreApplication resulted in QStandardPaths looking in
'/usr/share/picmi/picmi/', which doesn't exist. Leaving the organization
name unset fixed this issue, and exposed another problem with our
locate() usage in which we incorrectly appended the directory name to
the returned path.