Fix homepage link: this time the real fix
Hit a git pitfall with the previous commit. Sorry for that.

The hyperlink in >Help >About Picmi does not work, because the hyperlink
is put in the place of the additional free form text for that dialog.

To keep the parameter list of the constructor invocation short, do not
fix the constructor call, but remove the homepage from there and set the
homepage link afterwards in a separate call to setHomepage().

REVIEW: 125992